requirements-docker.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
uvloop>=0.19.0
httptools>=0.6.0

# API clients
openai>=1.99.2
elevenlabs>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Environment and utilities
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2

# Data storage
redis>=4.0.0
boto3>=1.34.0
google-cloud-firestore>=2.11.0

# Audio processing (minimal)
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Video processing (minimal)
moviepy>=1.0.3
